Implement recognise() and add data for numbers
The data is where to sample for each number. Keep in mind that it must be unique, which is rather hard to do for the 0-number since it needs to find a pixel which is always occupied by every number in the game because the background colour is the same.